components moved from collections >> collection for dynamic routing seeKennel for details
