+++ src/components/Collection/Has/HasCard.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react'
-import Dashboard from '../Dashboard'
+import Dashboard from '../../Dashboard/Dashboard'
 import './HasCard.css'
 
 class HasCard extends Component {
